Restored PDFParser, code style fixes and now no try!
